requirements.txt
feedparser
flask
Mastodon.py
python-dotenv
